backend/requirements.txt
fastapi
uvicorn
requests
google-cloud-aiplatform
google-cloud-secret-manager
google-cloud-logging
google-cloud-bigquery
PyGithub
python-dotenv 